+++ app/controllers/configs.py
@@ -24,3 +24,6 @@ class Config:
     DB_NAME = os.getenv('DB_NAME')
     DB_USERNAME = os.getenv('DB_USERNAME')
     DB_PASSWORD = os.getenv('DB_PASSWORD')
+
+    # API port, default to 5000
+    API_PORT = int(os.getenv('API_PORT', '5000'))