+++ app/controllers/pdl.py
@@ -69,8 +69,6 @@ HOUSE_PREDICTED_DATA_JSON = "house_outlay_max.json"
 HOUSE_PREDICTED_PRACTICE_CATEGORIES_DATA_JSON = "house_outlay_practices.json"
 TITLE_IV_DATA_PATH = os.path.join("controllers", "data", "title-iv")
 IV_SNAP_DATA_PATH = os.path.join(TITLE_IV_DATA_PATH, "programs", "snap")
-SNAP_STATE_DISTRIBUTION_DATA_JSON = "snap_state_distribution_data.json"
-SNAP_SUBPROGRAMS_DATA_JSON = "summary.json"
 
 TITLE_XI_DATA_PATH = os.path.join("controllers", "data", "title-xi")
 XI_CROP_INS_DATA_PATH = os.path.join(TITLE_XI_DATA_PATH, "programs", "crop-insurance")
@@ -758,7 +756,6 @@ def titles_title_xi_programs_crop_insurance_summary_search():
     endpoint_response = generate_title_xi_summary_response(program_id, start_year, end_year)
     return endpoint_response
 
-
 # /pdl/titles/title-iv/programs/snap/state-distribution
 def titles_title_iv_programs_snap_state_distribution_search():
     program_id = get_program_id(TITLE_IV_SNAP_PROGRAM_NAME)
@@ -774,6 +771,21 @@ def titles_title_iv_programs_snap_state_distribution_search():
     endpoint_response = generate_title_iv_state_distribution_response(program_id, start_year, end_year)
     return endpoint_response
 
+# /pdl/titles/title-iv/programs/snap/summary
+def titles_title_iv_programs_snap_summary_search():
+    program_id = get_program_id(TITLE_IV_SNAP_PROGRAM_NAME)
+    if program_id is None:
+        msg = {
+            "reason": "No record for the given program name " + TITLE_IV_SNAP_PROGRAM_NAME,
+            "error": "Not found: " + request.url,
+        }
+        logging.error("SNAP: " + json.dumps(msg))
+        return rs_handlers.not_found(msg)
+
+    start_year = 2018
+    end_year = 2022
+    endpoint_response = generate_title_iv_summary_response(program_id, start_year, end_year)
+    return endpoint_response
 
 # construct state
 def construct_state_result(state):
@@ -1167,6 +1179,30 @@ def generate_title_iv_state_distribution_response(program_id, start_year, end_ye
 
     return program_response_dict
 
+def generate_title_iv_summary_response(program_id, start_year, end_year):
+    session = Session()
+
+    # Construct the query
+    program_query = session.query(
+        func.avg(Payment.recipient_count).label('averageMonthlyRecipientCount'),
+        func.sum(Payment.payment).label('totalPaymentInDollars')
+    ).join(
+        Program, Payment.program_id == Program.id
+    ).filter(
+        Payment.program_id == program_id,
+        Payment.year.between(start_year, end_year)
+    )
+
+    # execute the query
+    result = program_query.first()
+
+    # aggregate dictionary for summing
+    aggregate_dict = {
+        'totalPaymentInDollars': round(result.totalPaymentInDollars, 2),
+        'averageMonthlyParticipation': round(result.averageMonthlyRecipientCount, 2)
+    }
+
+    return aggregate_dict
 
 def generate_title_i_total_state_distribution_response(title_id, start_year, end_year):
     session = Session()