+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add back summary endpoint to SNAP. [#271](https://github.com/policy-design-lab/pdl-api/issues/271)
 - Add top-level endpoints to Title-II. [#270](https://github.com/policy-design-lab/pdl-api/issues/270)
 
 ### Changed
 - Change house api path. [#295](https://github.com/policy-design-lab/pdl-api/issues/295)
+- Add feature to specify running port. [285](https://github.com/policy-design-lab/pdl-api/issues/285)
 
 ### Fixed
 - Fix calculation error in house outlay JSON data file. [#306](https://github.com/policy-design-lab/pdl-api/issues/306)